refactor!: Consolidate Query and QueryBuilder.
