+++ src/index.js
@@ -5,14 +5,12 @@ import Reducer from './engine/reduce/reducer';
 import parse from './expression/parse';
 import walk_ast from './expression/ast/walk';
 import Query from './query/query';
-import QueryBuilder from './query/query-builder';
 import { Verbs } from './query/verb';
 
 export const internal = {
   Table,
   ColumnTable,
   Query,
-  QueryBuilder,
   Reducer,
   Verbs,
   parse,
@@ -25,6 +23,6 @@ export { seed } from './util/random';
 export { default as fromArrow } from './format/from-arrow';
 export { default as fromCSV } from './format/from-csv';
 export { default as fromJSON } from './format/from-json';
-export { query, queryFrom } from './query/query-builder';
+export { query, queryFrom } from './query/query';
 export * from './op/register';
 export * from './verbs';
