+++ src/query/query.js
@@ -1,10 +1,35 @@
+import Transformable from '../table/transformable';
 import { Query as QueryType } from './constants';
-import { Verb } from './verb';
+import { Verb, Verbs } from './verb';
+
+/**
+ * Create a new query instance. The query interface provides
+ * a table-like verb API to construct a query that can be
+ * serialized or evaluated against Arquero tables.
+ * @param {string} [tableName] The name of the table to query. If
+ *  provided, will be used as the default input table to pull from
+ *  a provided catalog to run the query against.
+ * @return {Query} A new builder instance.
+ */
+export function query(tableName) {
+  return new Query(null, null, tableName);
+}
+
+/**
+ * Create a new query instance from a serialized object.
+ * @param {object} object A serialized query representation, such as
+ *  those generated by query(...).toObject().
+ * @returns {Query} The instantiated query instance.
+ */
+export function queryFrom(object) {
+  return Query.from(object);
+}
 
 /**
  * Model a query as a collection of serializble verbs.
+ * Provides a table-like interface for constructing queries.
  */
-export default class Query {
+export default class Query extends Transformable {
 
   /**
    * Construct a new query instance.
@@ -14,8 +39,8 @@ export default class Query {
    * @param {string} [table] Optional name of the table to query.
    */
   constructor(verbs, params, table) {
+    super(params);
     this._verbs = verbs || [];
-    this._params = params || null;
     this._table = table;
   }
 
@@ -27,6 +52,15 @@ export default class Query {
    */
   static from({ verbs, table, params }) {
     return new Query(verbs.map(Verb.from), params, table);
+  }
+
+  /**
+   * Provide an informative object string tag.
+   */
+  get [Symbol.toStringTag]() {
+    if (!this._verbs) return 'Object'; // bail if called on prototype
+    const ns = this._verbs.length;
+    return `Query: ${ns} verbs` + (this._table ? ` on '${this._table}'` : '');
   }
 
   /**
@@ -118,4 +152,14 @@ function serialize(query, method, props) {
     ...(query._params ? { params: query._params } : null),
     ...(query._table ? { table: query._table } : null)
   };
+}
+
+// Internal verb handlers
+for (const name in Verbs) {
+  const verb = Verbs[name];
+  Query.prototype['__' + name] = (qb, ...args) => new Query(
+    qb._verbs.concat(verb(...args)),
+    qb._params,
+    qb._table
+  );
 }
